seralize CallManyParams
